+++ src/admin/apis/NavController.php
@@ -146,7 +146,15 @@ class NavController extends \luya\admin\base\RestController
 
     public function actionGetProperties($navId)
     {
-        UserOnline::lock(Yii::$app->adminuser->id, NavItem::tableName(), $navId, 'lock_cms_edit_page', ['title' => Nav::findOne($navId)->activeLanguageItem->title]);
+        $nav = Nav::findOne($navId);
+
+        if (!$nav) {
+            throw new NotFoundHttpException();
+        }
+
+        UserOnline::lock(Yii::$app->adminuser->id, NavItem::tableName(), $navId, 'lock_cms_edit_page', [
+            'title' => $nav->defaultLanguageItem ? $nav->defaultLanguageItem->title : '(no translation)',
+        ]);
         
         $data = [];
         foreach (Property::find()->select(['admin_prop_id', 'value'])->where(['nav_id' => $navId])->asArray()->all() as $row) {