+++ tests/src/models/NavItemTest.php
@@ -2,13 +2,18 @@
 
 namespace tests\web\cmsadmin\models;
 
-use cmstests\CmsFrontendTestCase;
+use cmstests\WebModelTestCase;
+use luya\cms\models\Nav;
 use luya\cms\models\NavItem;
+use luya\testsuite\traits\CmsDatabaseTableTrait;
 
-class NavItemTest extends CmsFrontendTestCase
+class NavItemTest extends WebModelTestCase
 {
+    use CmsDatabaseTableTrait;
+
     public function testSlugifyAlias()
     {
+        $this->createCmsNavItemFixture();
         $model = new NavItem();
 
         $model->alias = "äÄöÖüÜß<>";
@@ -22,5 +27,38 @@ class NavItemTest extends CmsFrontendTestCase
         $model->alias = "a1Zあ新~!@#$^&*()_[];',:?";
         $model->slugifyAlias();
         $this->assertSame("a1zあ新~!@#$^&*()[];',:?", $model->alias);
+    }
+
+    public function testGetDefaultLanguageRelationGetter()
+    {
+        $this->createCmsNavFixture([
+            1  => [
+                'id' => 1,
+            ]
+        ]);
+        
+        $this->createCmsNavItemFixture([
+            1 => [
+                'id' => 1,
+                'nav_id' => 1,
+                'lang_id' => 1,
+            ]
+        ]);
+
+        $this->createAdminLangFixture([
+            1  => [
+                'id' => 1,
+                'name' => 'English',
+                'short_code' => 'en',
+                'is_default' => 1,
+                'is_deleted' => 0,
+            ]
+        ]);
+        
+        $nav = Nav::findOne(1);
+
+        $this->assertSame(1, $nav->id);
+        $this->assertSame(1, $nav->defaultLanguageItem->id);
+        $this->assertSame(1, $nav->activeLanguageItem->id);
     }
 }